fix : ranking job query && refactor : reflect feedback
